pyproject.toml
# PEP-518 https://peps.python.org/pep-0518/
[build-system]
# Minimum requirements for the build system to execute.
requires = ["setuptools>=69.1.1"]  # PEP 508 specifications.
build-backend = "setuptools.build_meta"

[project]
name = "noachdm"
authors = [
    {name = "NOA - Beyond"}
]
description = "Change Detection Mapping EO Processor"
requires-python = ">=3.11.14"
dependencies = [
    "build>=1.1.1",
    "click>=8.1.7",
    "dask[array]>=2024.1.0",
    "einops>=0.8.1",
    "geopandas~=1.1.0",
    "kafka-python-ng>=2.2.3",
    "requests-aws4auth~=1.3.1",
    "rioxarray>=0.19.0",
    "numcodecs>=0.12.1",
    "psycopg>=3.2.3",
    "shapely~=2.1.1",
    "torch~=2.7.1",
    "numcodecs~=0.12.1",
    "zarr~=2.16.1"
]
keywords = [
    "data pipelines",
    "data engineering",
]
license = {text = "AGPL-3.0"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.11.14",
]
dynamic = ["readme", "version"]

[project.optional-dependencies]
test = [
    "black>=24.2.0",
    "coverage>=7.4.3",
    "flake8>=7.0.0",
    "pytest>=8.1.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
noachdm = "noachdm:cli"

[tool.setuptools.packages.find]
include = ["noachdm*"]

[tool.setuptools.dynamic]
readme = {file = "README.md", content-type = "text/markdown"}
version = {attr = "noachdm.__version__"}

[tool.pytest.ini_options]
addopts="""
--cov-context test  \
--cov-config pyproject.toml \
--cov-report xml:coverage.xml \
--cov-report term-missing \
--cov noachdm \
--cov tests \
--no-cov-on-fail \
-ra \
-W ignore"""
testpaths = [
  "tests"
]